2016_11_04_153154_create_content_server_stats_table.php edited online with Bitbucket - Jim wants deploy_id to be a bigint.

+++ database/migrations/2016_11_04_153154_create_content_server_stats_table.php
@@ -15,7 +15,7 @@ class CreateContentServerStatsTable extends Migration
         Schema::create('content_server_stats', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->bigInteger('email_id')->unsigned()->default(0);
-            $table->integer('deploy_id')->unsigned()->default(0);
+            $table->bigInteger('deploy_id')->unsigned()->default(0);
             $table->integer('action_id')->unsigned()->default(0);
             $table->datetime('datetime');
             $table->timestamp('created_at')->default(DB::raw('CURRENT_TIMESTAMP'));
